maintenance  save api

+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/dto/DecokePlanningDTO.java
@@ -18,15 +18,15 @@ public class DecokePlanningDTO {
 
 	private UUID id;
     private String monthName;
-    private Integer ibr;
-    private Integer mnt;
-    private Integer shutdown;
-    private Integer sad;
-    private Integer totalSAD;
-    private Integer bbu;
-    private Integer demoHSS;
-    private Integer demoBBU;
-    private Integer demoSAD;
+    private Double ibr;
+    private Double mnt;
+    private Double shutdown;
+    private Double sad;
+    private Double totalSAD;
+    private Double bbu;
+    private Double demoHSS;
+    private Double demoBBU;
+    private Double demoSAD;
     private Double fourFD;
     private Double fourF;
     private Double fiveF;
@@ -35,9 +35,9 @@ public class DecokePlanningDTO {
     private String aopYear;
     private UUID plantId;
     private String remarks;
-    private Integer slowdown;
-    private Integer bbd;
-    private Integer demoSD;
-    private Integer numberOfDays;
-    private Integer coilReplacement;	    
+    private Double slowdown;
+    private Double bbd;
+    private Double demoSD;
+    private Double numberOfDays;
+    private Double coilReplacement;	    
 }
